refactor(app): extract Victory screen section widgets (#4224)
Decompose victory_screen_body.dart into focused sibling libraries
(section card, conditions, standings, standing rows, minimap layout).
Structural-only; GAME70001 keys and layout unchanged.

Refs #4224
