+++ app/lib/features/game/screens/victory/victory_section_card.dart
@@ -1,0 +1,40 @@
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+import 'package:flutter/material.dart';
+
+/// Bordered section chrome for Victory panel blocks (conditions, standings, minimap).
+class VictorySectionCard extends StatelessWidget {
+  const VictorySectionCard({
+    super.key,
+    required this.title,
+    required this.child,
+  });
+
+  final String title;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return DecoratedBox(
+      decoration: BoxDecoration(
+        color: EditorialMonoclePalette.surface,
+        border: Border.all(color: EditorialMonoclePalette.border),
+      ),
+      child: Padding(
+        padding: const EdgeInsets.all(10),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              title,
+              style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                color: EditorialMonoclePalette.fg,
+              ),
+            ),
+            const SizedBox(height: 8),
+            child,
+          ],
+        ),
+      ),
+    );
+  }
+}
